refactor: unify event records source code location
